+++ HTTPShortcuts/app/src/main/kotlin/ch/rmy/android/http_shortcuts/activities/categories/editor/CategoryEditorContent.kt
@@ -219,10 +219,10 @@ private fun BackgroundColorButton(
         modifier = Modifier
             .fillMaxWidth()
             .background(Color(backgroundColor))
-            .padding(Spacing.SMALL)
             .clickable {
                 onColorButtonClicked()
-            },
+            }
+            .padding(Spacing.SMALL),
     )
 }