+++ HTTPShortcuts/app/src/main/kotlin/ch/rmy/android/http_shortcuts/activities/variablewidget/VariableWidgetSettingsScreen.kt
@@ -45,11 +45,13 @@ fun VariableWidgetSettingsScreen(
             variableValue = viewState.variableValue,
             fontSize = viewState.fontSize,
             title = viewState.title,
+            background = viewState.background,
             shortcutId = viewState.shortcutId,
             shortcuts = viewState.selectableShortcuts,
             onVariableSelected = viewModel::onVariableSelected,
             onFontSizeChanged = viewModel::onFontSizeChanged,
             onTitleChanged = viewModel::onTitleChanged,
+            onBackgroundChanged = viewModel::onBackgroundChanged,
             onShortcutSelected = viewModel::onShortcutSelected,
         )
     }