+++ HTTPShortcuts/app/src/main/kotlin/ch/rmy/android/http_shortcuts/activities/variablewidget/VariableWidgetSettingsViewState.kt
@@ -4,6 +4,7 @@ import androidx.compose.runtime.Stable
 import ch.rmy.android.http_shortcuts.activities.variablewidget.models.SelectableShortcut
 import ch.rmy.android.http_shortcuts.activities.variablewidget.models.SelectableVariable
 import ch.rmy.android.http_shortcuts.data.domains.shortcuts.ShortcutId
+import ch.rmy.android.http_shortcuts.data.enums.WidgetBackgroundType
 
 @Stable
 data class VariableWidgetSettingsViewState(
@@ -12,6 +13,7 @@ data class VariableWidgetSettingsViewState(
     val variableValue: String?,
     val fontSize: Int,
     val title: String,
+    val background: WidgetBackgroundType?,
     val shortcutId: ShortcutId?,
     val selectableShortcuts: List<SelectableShortcut>,
 ) {