+++ HTTPShortcuts/app/src/main/kotlin/ch/rmy/android/http_shortcuts/data/domains/widgets/VariableWidgetsRepository.kt
@@ -4,6 +4,7 @@ import ch.rmy.android.http_shortcuts.data.Database
 import ch.rmy.android.http_shortcuts.data.domains.BaseRepository
 import ch.rmy.android.http_shortcuts.data.domains.shortcuts.ShortcutId
 import ch.rmy.android.http_shortcuts.data.domains.variables.GlobalVariableId
+import ch.rmy.android.http_shortcuts.data.enums.WidgetBackgroundType
 import ch.rmy.android.http_shortcuts.data.models.VariableWidget
 import javax.inject.Inject
 
@@ -17,6 +18,7 @@ constructor(
         variableId: GlobalVariableId,
         fontSize: Int,
         title: String,
+        background: WidgetBackgroundType?,
         shortcutId: ShortcutId?,
     ) = query {
         variableWidgetDao().insertOrReplace(
@@ -25,6 +27,7 @@ constructor(
                 variableId = variableId,
                 fontSize = fontSize,
                 title = title,
+                background = background,
                 shortcutId = shortcutId,
             ),
         )