+++ HTTPShortcuts/app/src/main/kotlin/ch/rmy/android/http_shortcuts/data/models/VariableWidget.kt
@@ -5,6 +5,7 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 import ch.rmy.android.http_shortcuts.data.domains.shortcuts.ShortcutId
 import ch.rmy.android.http_shortcuts.data.domains.variables.GlobalVariableId
+import ch.rmy.android.http_shortcuts.data.enums.WidgetBackgroundType
 
 @Entity(tableName = "variable_widget")
 data class VariableWidget(
@@ -19,4 +20,6 @@ data class VariableWidget(
     val title: String,
     @ColumnInfo(name = "shortcut_id")
     val shortcutId: ShortcutId? = null,
+    @ColumnInfo(name = "background")
+    val background: WidgetBackgroundType? = null,
 )